Removed __version__ import to setup.py and moved signal into signals.py
