+++ django_guid/__init__.py
@@ -3,4 +3,6 @@ Don't put imports or code here.
 This file is imported by setup.py
 Adding imports here will break setup.py
 """
-__version__ = '1.1.1'
+from . import signals  # noqa: F401
+
+__version__ = '1.2.0'
