+++ django_guid/middleware.py
@@ -1,10 +1,8 @@
 import logging
 import threading
 import uuid
-from typing import Callable, Optional, Union
+from typing import Callable, Union
 
-from django.core.signals import request_finished
-from django.dispatch import receiver
 from django.http import HttpRequest, HttpResponse
 
 from .config import settings
@@ -143,17 +141,3 @@ class GuidMiddleware(object):
             )
 
         return request.correlation_id
-
-
-@receiver(request_finished)
-def delete_guid(sender: Optional[dict], **kwargs: dict) -> None:
-    """
-    Receiver function for when a request finishes.
-    When a request is finished,
-    we make sure that the current requests _guid reference is deleted to prevent a memory leak.
-
-    :param sender: dict or None
-    :param kwargs: dict
-    :return: None
-    """
-    GuidMiddleware.delete_guid()
